Move `FeedViewAdapter` to a separate file
